fix(webauthn): Hardening and type corrections
- Return StateNotFound when no state rows are deleted (replay
  prevention)
- Map new error variants (CounterVerification, CredentialNotFound,
  StateNotFound)
- Fix counter column type i32 → u32 (WebAuthn counters are unsigned)
- Allow description to be cleared on credential update
- Clean up register/finish.rs early-return structure

+++ crates/webauthn/src/driver/sql/model/webauthn_credential.rs
@@ -24,7 +24,7 @@ pub struct Model {
     pub description: Option<String>,
     #[sea_orm(column_type = "Text")]
     pub passkey: String,
-    pub counter: i32,
+    pub counter: i64,
     pub r#type: String,
     pub aaguid: Option<String>,
     pub created_at: DateTime,
